fix: Reuse actions for steps
